Run extras
