+++ core/src/main/kotlin/org/sourcegrade/jagr/core/CommonModule.kt
@@ -28,10 +28,12 @@ import org.sourcegrade.jagr.core.executor.*
 import org.sourcegrade.jagr.core.export.rubric.*
 import org.sourcegrade.jagr.core.export.submission.EclipseSubmissionExporter
 import org.sourcegrade.jagr.core.export.submission.GradleSubmissionExporter
+import org.sourcegrade.jagr.core.extra.ExtrasManagerImpl
 import org.sourcegrade.jagr.core.io.*
 import org.sourcegrade.jagr.core.rubric.*
 import org.sourcegrade.jagr.core.rubric.grader.*
 import org.sourcegrade.jagr.core.testing.*
+import org.sourcegrade.jagr.launcher.io.ExtrasManager
 import org.sourcegrade.jagr.launcher.env.*
 import org.sourcegrade.jagr.launcher.executor.*
 import org.sourcegrade.jagr.launcher.io.*
@@ -49,6 +51,7 @@ class CommonModule(private val configuration: LaunchConfiguration) : AbstractMod
     bind(GradingQueue.Factory::class.java).to(GradingQueueFactoryImpl::class.java)
     bind(RuntimeGrader::class.java).to(RuntimeGraderImpl::class.java)
     bind(SerializerFactory.Locator::class.java).to(SerializationFactoryLocatorImpl::class.java)
+    bind(ExtrasManager::class.java).to(ExtrasManagerImpl::class.java)
 
     bind(Criterion.Factory::class.java).to(CriterionFactoryImpl::class.java)
     bind(CriterionHolderPointCalculator.Factory::class.java).to(CriterionHolderPointCalculatorFactoryImpl::class.java)
